# api/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "code": {
            "description": "HEIMDALL diagnostic code.",
            "pattern": "^HD[0-9]{4}$",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "reason_code": {
            "description": "SESAME's decision reason, present on 403.",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "sessionBearer": {
        "description": "Authorization: Bearer <session-id>.<session-secret>, issued by SESAME. HEIMDALL neither mints nor verifies it locally; it forwards it to the engine.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The control-plane API. It is the contract the CLI, CI, and Terraform use; the Tachyon web tier is a thin proxy in front of it and adds no routes.\n\nEvery route except /healthz, /readyz, and /api/v1/version resolves to exactly one SESAME authorization action, listed as x-heimdall-action, evaluated once in middleware before any handler runs. A 403 carries SESAME's reason_code verbatim; a 503 means the engine did not answer and is never a bypass.\n\nGenerated from the route table by `heimdall contract`. Do not hand-edit.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "HEIMDALL control plane API",
    "version": "26.33.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audit": {
      "get": {
        "operationId": "get_api_v1_audit",
        "responses": {
          "200": {
            "description": "Authorized."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource identifier is not well formed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Denied by SESAME. Carries reason_code."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such application."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The sync completed with per-service failures."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The compose file uses something this provider cannot express; the message names the offending service."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authorization engine unavailable. Never a bypass."
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "summary": "audit:read on tenant:self",
        "x-heimdall-action": "audit:read",
        "x-heimdall-mutating": false,
        "x-heimdall-resource": "tenant:self"
      }
    },
    "/api/v1/audit/export": {
      "get": {
        "operationId": "get_api_v1_audit_export",
        "responses": {
          "200": {
            "description": "Authorized."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource identifier is not well formed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Denied by SESAME. Carries reason_code."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such application."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The sync completed with per-service failures."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The compose file uses something this provider cannot express; the message names the offending service."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authorization engine unavailable. Never a bypass."
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "summary": "audit:export on tenant:self",
        "x-heimdall-action": "audit:export",
        "x-heimdall-mutating": false,
        "x-heimdall-resource": "tenant:self"
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "get_api_v1_projects",
        "responses": {
          "200": {
            "description": "Authorized."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource identifier is not well formed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Denied by SESAME. Carries reason_code."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such application."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The sync completed with per-service failures."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The compose file uses something this provider cannot express; the message names the offending service."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authorization engine unavailable. Never a bypass."
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "summary": "project:read on tenant:self",
        "x-heimdall-action": "project:read",
        "x-heimdall-mutating": false,
        "x-heimdall-resource": "tenant:self"
      }
    },
    "/api/v1/projects/{project}/apps": {
      "get": {
        "operationId": "get_api_v1_projects_project_apps",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9._-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource identifier is not well formed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Denied by SESAME. Carries reason_code."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such application."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The sync completed with per-service failures."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The compose file uses something this provider cannot express; the message names the offending service."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authorization engine unavailable. Never a bypass."
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "summary": "app:read on project:{project}",
        "x-heimdall-action": "app:read",
        "x-heimdall-mutating": false,
        "x-heimdall-resource": "project:{project}"
      },
      "post": {
        "operationId": "post_api_v1_projects_project_apps",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9._-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource identifier is not well formed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Denied by SESAME. Carries reason_code."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such application."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The sync completed with per-service failures."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The compose file uses something this provider cannot express; the message names the offending service."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authorization engine unavailable. Never a bypass."
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "summary": "app:create on project:{project}",
        "x-heimdall-action": "app:create",
        "x-heimdall-mutating": true,
        "x-heimdall-resource": "project:{project}"
      }
    },
...
    "/api/v1/projects/{project}/apps/{app}/events": {
      "get": {
        "operationId": "get_api_v1_projects_project_apps_app_events",
        "parameters": [
          {
            "in": "path",
            "name": "app",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9._-]+$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9._-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource identifier is not well formed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Denied by SESAME. Carries reason_code."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such application."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The sync completed with per-service failures."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The compose file uses something this provider cannot express; the message names the offending service."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authorization engine unavailable. Never a bypass."
          }
        },
        "security": [
          {
            "sessionBearer": []
          }
        ],
        "summary": "observe:events on project:{project}:app:{app}",
        "x-heimdall-action": "observe:events",
        "x-heimdall-mutating": false,
        "x-heimdall-resource": "project:{project}:app:{app}"
      }
    },
...
  },
  "servers": [
    {
      "description": "Default local control plane",
      "url": "http://127.0.0.1:8080"
    }
  ]
}